+++ src/cli.js
@@ -22,7 +22,7 @@ import {
 } from './saihubot-cli-skill-eth';
 import {skillEth2Stats, skillBeaconLastBlock, skillGetValidatorBlance} from './saihubot-cli-skill-eth2';
 import {skillsGas} from './saihubot-cli-skill-ethgas';
-import {skillGetBlance} from './saihubot-cli-skill-balance';
+import {skillGetBlance, skillGetEthBlance} from './saihubot-cli-skill-balance';
 import {skills as skillsXDAI} from './saihubot-cli-skill-xdai';
 import {skillChainId} from './saihubot-cli-skill-chainid';
 import {skillAwesomeEth} from './saihubot-skill-awesome';
@@ -59,6 +59,7 @@ const bot = new SaihuBot({
   ],
   skills: [
     skillGetBlance,
+    skillGetEthBlance,
     skillLastBlock,
     ...skillsAddress,
     skillGetValidatorBlance,